+++ pi4-app/src/main/java/br/senac/service/impl/ConcessionariaServiceImpl.java
@@ -31,7 +31,7 @@ public class ConcessionariaServiceImpl implements ConcessionariaService{
 		
 	}
 
-	public Concessionaria getConssecionaria(Integer id) {
+	public Concessionaria getConssecionaria(Long id) {
 		return concessionariaEAO.getConssecionaria(id);
 	}
